+++ src/autocomplete-constants.ts
@@ -12,9 +12,7 @@ export const KEYCODES = {
     PAGEDOWN: 34,
     END: 35,
     HOME: 36,
-    LEFT: 37,
     UP: 38,
-    RIGHT: 39,
     DOWN: 40,
     DELETE: 46,
 };